# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'ModeMakers API'
  description: 'documentation for modeMakers api'
  version: 1.0.0
servers:
  -
    url: /docs
paths:
...
  /api/review:
    get:
      summary: ''
      operationId: getApiReview
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    post:
      summary: ''
      operationId: postApiReview
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the users table.'
                  example: pariatur
                  nullable: false
                product_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the products table.'
                  example: quod
                  nullable: false
                comment:
                  type: string
                  description: ''
                  example: illo
                  nullable: false
              required:
                - user_id
                - product_id
                - comment
      security: []
  '/api/review/{id}':
    get:
      summary: ''
      operationId: getApiReviewId
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    put:
      summary: ''
      operationId: putApiReviewId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the users table.'
                  example: null
                  nullable: false
                product_id:
                  type: string
                  description: 'The <code>id</code> of an existing record in the products table.'
                  example: null
                  nullable: false
                comment:
                  type: string
                  description: ''
                  example: laboriosam
                  nullable: false
      security: []
    delete:
      summary: ''
      operationId: deleteApiReviewId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the review.'
        example: aliquam
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []